fix: cleanup empty metadata to fix variable creation (#9529)
